feat: Move MediaUploader to components
