+++ imports/plugins/core/ui/client/components/media/index.js
@@ -1,2 +1,3 @@
 export { default as MediaGallery } from "./mediaGallery";
-export { default as MediaItem } from "./media";
+export { default as MediaItem } from "./mediaItem";
+export { default as MediaUploader } from "./mediaUploader";
